+++ tests/unit/test_client.py
@@ -933,6 +933,19 @@ class TestClientMethods(unittest.TestCase):
         self.client.set_default_interchain_network("a", "b")
         self.client.request.post.assert_called_once_with("/v1/interchains/default", {"version": "1", "blockchain": "a", "name": "b"})
 
+    def test_publish_interchain_transaction_throws_correctly(self, mock_creds, mock_request):
+        self.client = dragonchain_sdk.create_client()
+        self.assertRaises(TypeError, self.client.publish_interchain_transaction, 1, "b", "c")
+        self.assertRaises(TypeError, self.client.publish_interchain_transaction, "a", 2, "c")
+        self.assertRaises(TypeError, self.client.publish_interchain_transaction, "a", "b", 3)
+
+    def test_publish_interchain_transaction_calls_correctly(self, mock_creds, mock_request):
+        self.client = dragonchain_sdk.create_client()
+        -self.client.publish_interchain_transaction("a", "b", "c")
+        self.client.request.post.assert_called_once_with(
+            "/v1/interchains/transaction/publish", {"version": "1", "blockchain": "a", "name": "b", "signed_txn": "c"}
+        )
+
     def test_list_interchain_throws_correctly(self, mock_creds, mock_request):
         self.client = dragonchain_sdk.create_client()
         self.assertRaises(TypeError, self.client.list_interchain_networks, 1)